+++ src/pydiverse/colspec/_rule.py
@@ -124,7 +124,7 @@ def rule_polars(
     """
 
     def decorator(validation_fn: ValidationFunctionPolars) -> RulePolars:
-        if pl.Expr is None:
+        if pl.Expr is object:
             # decorator should also work if polars is not installed
             if group_by is not None:
                 return GroupRulePolars(expr=None, group_columns=group_by)